added initial behavior of game BreakOut
