+++ Break_Out_Game/Break_Out_Game/Break_Out_Game.cpp
@@ -2,12 +2,25 @@
 //
 
 #include "Break_Out_Game.h"
+#include "Breakout.h"
 #include "olcConsoleGameEngine.h"
+//#include ""
+//using namespace std;
 
-using namespace std;
+static const int Levels = 3;
+static const int Lives = 5;
+
+
 
 int main()
 {
-	cout << "Hello CMake." << endl;
+	std::cout << "Hello CMake." << std::endl;
+
+	Brekout Game_Breakout;
+	Game_Breakout.ConstructConsole(256, 240, 4, 4);
+	Game_Breakout.Start();
+
+
+
 	return 0;
 }
