+++ Break_Out_Game/Break_Out_Game/Breakout.cpp
@@ -1,0 +1,13 @@
+#include "Breakout.h"
+
+bool Brekout::OnUserCreate() {
+	//	
+
+	return true;
+}
+
+bool Brekout::OnUserUpdate(float fElapsedTime) {
+
+
+	return true;
+}
