+++ Break_Out_Game/Break_Out_Game/Breakout.h
@@ -1,0 +1,17 @@
+#include "olcConsoleGameEngine.h"
+#ifndef BREAKOUT
+#define BREAKOUT
+
+class Brekout : public olcConsoleGameEngine {
+
+public:
+
+public:
+	Brekout() {}
+	~Brekout() {}
+public:
+	bool OnUserCreate();
+	bool OnUserUpdate(float fElapsedTime);
+};
+
+#endif
