+++ Break_Out_Game/Break_Out_Game/CMakeLists.txt
@@ -7,6 +7,6 @@ cmake_minimum_required (VERSION 3.8)
 add_definitions(-DUNICODE -D_UNICODE)
 
 # Добавьте источник в исполняемый файл этого проекта.
-add_executable (Break_Out_Game "Break_Out_Game.cpp" "Break_Out_Game.h" "olcConsoleGameEngine.h")
+add_executable (Break_Out_Game "Break_Out_Game.cpp" "Break_Out_Game.h" "olcConsoleGameEngine.h" "Breakout.cpp" "Breakout.h" "olcConsoleGameEngine.cpp" "olcSprite.h" "olcSprite.cpp" "engine_constants.h")
 
 # TODO: Добавьте тесты и целевые объекты, если это необходимо.
